Unfortunately, u32 is clearly not enough to avoid collisions in a codebase as large as Core. This PR reverts the ID sizes back to 64 bits and removes tagging.

Here are the unsuccessful things I tried to avoid collision with u32:

- Switching our ID hashing from strings to bytes (this does have better performance as it avoids string allocations, but doesn't help with collisions)
- Partitioning parts of the IDs with more bits for data that better differentiates entities. For example, URI + offset is much higher relevance for constant references than name since you cannot put two references in the same exact spot in the code (only exception is hidden singleton references)
- Hashing the ID parts separately and then combining them
- Kind tagging

### Idea

The only ID that actually fits in u32 is `UriId`. One idea I had is that we could partition `DefinitionId` and `ReferenceId` to store the URI in it. That way, we don't pay the price of also storing the UriId in these structs and still avoid collisions. That's an experiment for a different PR though. Let's eliminate the collisions first.

+++ rust/rubydex-sys/src/reference_api.rs
@@ -18,13 +18,13 @@ pub enum ReferenceKind {
 /// Shared iterator over reference (id, kind) pairs
 #[derive(Debug)]
 pub struct ReferencesIter {
-    pub entries: Box<[(u32, ReferenceKind)]>,
+    pub entries: Box<[(u64, ReferenceKind)]>,
     pub index: usize,
 }
 
 impl ReferencesIter {
     #[must_use]
-    pub fn new(entries: Box<[(u32, ReferenceKind)]>) -> *mut ReferencesIter {
+    pub fn new(entries: Box<[(u64, ReferenceKind)]>) -> *mut ReferencesIter {
         Box::into_raw(Box::new(ReferencesIter { entries, index: 0 }))
     }
 }
@@ -54,7 +54,7 @@ pub unsafe extern "C" fn rdx_references_iter_len(iter: *const ReferencesIter) ->
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn rdx_references_iter_next(
     iter: *mut ReferencesIter,
-    out_id: *mut u32,
+    out_id: *mut u64,
     out_kind: *mut ReferenceKind,
 ) -> bool {
     if iter.is_null() || out_id.is_null() || out_kind.is_null() {
@@ -101,7 +101,7 @@ pub unsafe extern "C" fn rdx_references_iter_free(iter: *mut ReferencesIter) {
 ///
 /// This function will panic if the reference cannot be found.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_constant_reference_name(pointer: GraphPointer, reference_id: u32) -> *const c_char {
+pub unsafe extern "C" fn rdx_constant_reference_name(pointer: GraphPointer, reference_id: u64) -> *const c_char {
     with_graph(pointer, |graph| {
         let ref_id = ReferenceId::new(reference_id);
         let reference = graph.constant_references().get(&ref_id).expect("Reference not found");
@@ -127,7 +127,7 @@ pub unsafe extern "C" fn rdx_constant_reference_name(pointer: GraphPointer, refe
 ///
 /// This function will panic if the reference cannot be found.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_method_reference_name(pointer: GraphPointer, reference_id: u32) -> *const c_char {
+pub unsafe extern "C" fn rdx_method_reference_name(pointer: GraphPointer, reference_id: u64) -> *const c_char {
     with_graph(pointer, |graph| {
         let ref_id = ReferenceId::new(reference_id);
         let reference = graph.method_references().get(&ref_id).expect("Reference not found");
@@ -152,7 +152,7 @@ pub unsafe extern "C" fn rdx_method_reference_name(pointer: GraphPointer, refere
 ///
 /// This function will panic if a reference or document cannot be found.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_constant_reference_location(pointer: GraphPointer, reference_id: u32) -> *mut Location {
+pub unsafe extern "C" fn rdx_constant_reference_location(pointer: GraphPointer, reference_id: u64) -> *mut Location {
     with_graph(pointer, |graph| {
         let ref_id = ReferenceId::new(reference_id);
         let reference = graph.constant_references().get(&ref_id).expect("Reference not found");
@@ -177,7 +177,7 @@ pub unsafe extern "C" fn rdx_constant_reference_location(pointer: GraphPointer, 
 ///
 /// This function will panic if a reference or document cannot be found.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_method_reference_location(pointer: GraphPointer, reference_id: u32) -> *mut Location {
+pub unsafe extern "C" fn rdx_method_reference_location(pointer: GraphPointer, reference_id: u64) -> *mut Location {
     with_graph(pointer, |graph| {
         let ref_id = ReferenceId::new(reference_id);
         let reference = graph.method_references().get(&ref_id).expect("Reference not found");
